+++ packages/core/styles/src/index.ts
@@ -40,6 +40,7 @@ export type { BoxSprinkle } from "./packages/atomic/box";
 export type { IconButtonSprinkle } from "./packages/atomic/iconButton";
 export type { TextSprinkle } from "./packages/atomic/text";
 export type { PopoverSprinkle } from "./packages/atomic/popover";
+export type { TitleSprinkle } from "./packages/atomic/title";
 export type { ModalSprinkle } from "./packages/composite/modal";
 export type { SidebarSprinkle } from "./packages/composite/sidebar";
 export type { TableSprinkle } from "./packages/composite/table";